Tweaked auto stacking, added skills bot code
I added the skills robot code. I also redid the autostacking code so that the chainbar goes to a fixed position when running out (rather than running for a set time). This helps with pickimg up from the floor. The resting position is just above the tips of the cones/this makes the robot more manuiverable